repair_screenshot

+++ src/draw_tools/display.c
@@ -32,7 +32,7 @@ void			pixel_put(t_env *v, int x, int y, t_color color)
 	if (v->p.sc.filter == 0)
 		v->pixels[y * v->w + x] = get_hex_rgba(color.r * 255, color.g * 255, color.b * 255, color.a);
 	else if (v->p.sc.filter == 1)
-		v->pixels[y * v->w + x] = (Uint32)greyscale(get_hex_rgba(color.r * 255, color.g * 255, color.b * 255, color.a));
+		v->pixels[y * v->w + x] = greyscale(get_hex_rgba(color.r * 255, color.g * 255, color.b * 255, color.a));
 	else if (v->p.sc.filter == 2)
 		v->pixels[y * v->w + x] = sepia(get_hex_rgba(color.r * 255, color.g * 255, color.b * 255, color.a));
 	else if (v->p.sc.filter == 3)
